+++ crates/client/rpc-core/src/utils.rs
@@ -4,7 +4,6 @@ use anyhow::{anyhow, Result};
 use base64::engine::general_purpose;
 use base64::Engine;
 use cairo_vm::types::program::Program;
-use flate2::read::GzDecoder;
 use frame_support::inherent::BlockT;
 use mp_digest_log::find_starknet_block;
 use mp_starknet::block::Block as StarknetBlock;
@@ -16,9 +15,8 @@ use sp_api::HeaderT;
 use sp_blockchain::HeaderBackend;
 use sp_runtime::BoundedVec;
 use starknet_core::types::{
-    BroadcastedDeclareTransaction, BroadcastedDeployAccountTransaction, BroadcastedInvokeTransaction,
     BroadcastedTransaction, CompressedLegacyContractClass, ContractClass, FromByteArrayError, LegacyContractEntryPoint,
-    LegacyEntryPointsByType, StarknetError,
+    LegacyEntryPointsByType,
 };
 
 /// Returns a `ContractClass` from a `ContractClassWrapper`
@@ -65,125 +63,17 @@ pub(crate) fn encode_base64(data: &[u8]) -> String {
 /// * `Transaction` - The converted transaction
 pub fn to_tx(request: BroadcastedTransaction, chain_id: &str) -> Result<Transaction> {
     match request {
-        BroadcastedTransaction::Invoke(invoke_tx) => to_invoke_tx(invoke_tx).map(|inner| inner.from_invoke(chain_id)),
+        BroadcastedTransaction::Invoke(invoke_tx) => {
+            InvokeTransaction::try_from(invoke_tx).map(|inner| inner.from_invoke(chain_id))
+        }
         BroadcastedTransaction::Declare(declare_tx) => {
-            to_declare_tx(declare_tx).map(|inner| inner.from_declare(chain_id))
+            DeclareTransaction::try_from(declare_tx).map(|inner| inner.from_declare(chain_id))
         }
-        BroadcastedTransaction::DeployAccount(deploy_account_tx) => to_deploy_account_tx(deploy_account_tx)
-            .and_then(|inner| inner.from_deploy(chain_id).map_err(|e| anyhow!(e))),
-    }
-}
-
-/// Converts a broadcasted invoke transaction to an invoke transaction
-pub fn to_invoke_tx(tx: BroadcastedInvokeTransaction) -> Result<InvokeTransaction> {
-    match tx {
-        BroadcastedInvokeTransaction::V0(_) => Err(StarknetError::FailedToReceiveTransaction.into()),
-        BroadcastedInvokeTransaction::V1(invoke_tx_v1) => Ok(InvokeTransaction {
-            version: 1_u8,
-            signature: BoundedVec::try_from(
-                invoke_tx_v1.signature.iter().map(|x| (*x).into()).collect::<Vec<Felt252Wrapper>>(),
-            )
-            .map_err(|e| anyhow!("failed to convert signature: {:?}", e))?,
-
-            sender_address: invoke_tx_v1.sender_address.into(),
-            nonce: Felt252Wrapper::from(invoke_tx_v1.nonce),
-            calldata: BoundedVec::try_from(
-                invoke_tx_v1.calldata.iter().map(|x| (*x).into()).collect::<Vec<Felt252Wrapper>>(),
-            )
-            .map_err(|e| anyhow!("failed to convert calldata: {:?}", e))?,
-            max_fee: Felt252Wrapper::from(invoke_tx_v1.max_fee),
-        }),
-    }
-}
-
-/// Converts a broadcasted deploy account transaction to a deploy account transaction
-pub fn to_deploy_account_tx(tx: BroadcastedDeployAccountTransaction) -> Result<DeployAccountTransaction> {
-    let contract_address_salt = tx.contract_address_salt.into();
-
-    let account_class_hash = tx.class_hash;
-
-    let signature = tx
-        .signature
-        .iter()
-        .map(|f| (*f).into())
-        .collect::<Vec<Felt252Wrapper>>()
-        .try_into()
-        .map_err(|_| anyhow!("failed to bound signatures Vec<H256> by MaxArraySize"))?;
-
-    let calldata = tx
-        .constructor_calldata
-        .iter()
-        .map(|f| (*f).into())
-        .collect::<Vec<Felt252Wrapper>>()
-        .try_into()
-        .map_err(|_| anyhow!("failed to bound calldata Vec<U256> by MaxCalldataSize"))?;
-
-    let nonce = Felt252Wrapper::from(tx.nonce);
-    let max_fee = Felt252Wrapper::from(tx.max_fee);
-
-    Ok(DeployAccountTransaction {
-        version: 1_u8,
-        calldata,
-        salt: contract_address_salt,
-        signature,
-        account_class_hash: account_class_hash.into(),
-        nonce,
-        max_fee,
-    })
-}
-
-/// Converts a broadcasted declare transaction to a declare transaction
-pub fn to_declare_tx(tx: BroadcastedDeclareTransaction) -> Result<DeclareTransaction> {
-    match tx {
-        BroadcastedDeclareTransaction::V1(declare_tx_v1) => {
-            let signature = declare_tx_v1
-                .signature
-                .iter()
-                .map(|f| (*f).into())
-                .collect::<Vec<Felt252Wrapper>>()
-                .try_into()
-                .map_err(|_| anyhow!("failed to bound signatures Vec<H256> by MaxArraySize"))?;
-
-            // Create a GzipDecoder to decompress the bytes
-            let mut gz = GzDecoder::new(&declare_tx_v1.contract_class.program[..]);
-
-            // Read the decompressed bytes into a Vec<u8>
-            let mut decompressed_bytes = Vec::new();
-            std::io::Read::read_to_end(&mut gz, &mut decompressed_bytes)
-                .map_err(|_| anyhow!("Failed to decompress the contract class program"))?;
-
-            // Deserialize it then
-            let program: Program = Program::from_bytes(&decompressed_bytes, None)
-                .map_err(|_| anyhow!("Failed to deserialize the contract class program"))?;
-
-            Ok(DeclareTransaction {
-                version: 1_u8,
-                sender_address: declare_tx_v1.sender_address.into(),
-                nonce: Felt252Wrapper::from(declare_tx_v1.nonce),
-                max_fee: Felt252Wrapper::from(declare_tx_v1.max_fee),
-                signature,
-                contract_class: ContractClassWrapper {
-                    program: program.into(),
-                    entry_points_by_type: EntrypointMapWrapper::new(to_hash_map_entrypoints(
-                        declare_tx_v1.contract_class.entry_points_by_type.clone(),
-                    )),
-                },
-                compiled_class_hash: Felt252Wrapper::ZERO, // TODO: compute class hash
-            })
+        BroadcastedTransaction::DeployAccount(deploy_account_tx) => {
+            DeployAccountTransaction::try_from(deploy_account_tx)
+                .and_then(|inner| inner.from_deploy(chain_id).map_err(|e| anyhow!(e)))
         }
-        BroadcastedDeclareTransaction::V2(_) => Err(StarknetError::FailedToReceiveTransaction.into()),
     }
-}
-
-/// Returns a [HashMap<EntryPointTypeWrapper, Vec<EntryPointWrapper>>] from
-/// [LegacyEntryPointsByType]
-fn to_hash_map_entrypoints(entries: LegacyEntryPointsByType) -> HashMap<EntryPointTypeWrapper, Vec<EntryPointWrapper>> {
-    let mut entry_points_by_type = HashMap::default();
-
-    entry_points_by_type.insert(EntryPointTypeWrapper::Constructor, get_entrypoint_value(entries.constructor));
-    entry_points_by_type.insert(EntryPointTypeWrapper::External, get_entrypoint_value(entries.external));
-    entry_points_by_type.insert(EntryPointTypeWrapper::L1Handler, get_entrypoint_value(entries.l1_handler));
-    entry_points_by_type
 }
 
 /// Returns a [Result<LegacyEntryPointsByType>] (blockifier type) from a [EntrypointMapWrapper]
@@ -214,10 +104,6 @@ fn to_legacy_entry_points_by_type(entries: &EntrypointMapWrapper) -> Result<Lega
     Ok(LegacyEntryPointsByType { constructor, external, l1_handler })
 }
 
-/// Returns a [Vec<EntryPointWrapper>] from a [Vec<LegacyContractEntryPoint>]
-fn get_entrypoint_value(entries: Vec<LegacyContractEntryPoint>) -> Vec<EntryPointWrapper> {
-    entries.iter().map(|e| EntryPointWrapper::from(e.clone())).collect::<Vec<_>>()
-}
 /// Returns the current Starknet block from the block header's digest
 pub fn get_block_by_block_hash<B, C>(client: &C, block_hash: <B as BlockT>::Hash) -> Option<StarknetBlock>
 where
