+++ crates/primitives/starknet/Cargo.toml
@@ -45,8 +45,8 @@ serde_json = { version = "1.0.96", default-features = false }
 thiserror-no-std = { workspace = true }
 derive_more = { workspace = true, features = ["constructor"] }
 lazy_static = { workspace = true }
-# anyhow = { workspace = true }
-
+anyhow = { workspace = true, optional = true }
+flate2 = { workspace = true, optional = true }
 
 [dev-dependencies]
 rand = "0.8.5"
@@ -57,6 +57,8 @@ pretty_assertions = { workspace = true }
 [features]
 default = ["std"]
 std = [
+  "anyhow",
+  "flate2",
   "scale-codec/std",
   "scale-info/std",
   "bitvec/std",
